refactor(polarsdf): extract column_helpers and dataframe_helpers modules
Move module-level helper functions out of column.py and dataframe.py
into dedicated helper modules to keep the core class files slim.

column_helpers.py: schema context, dtype resolution, comparison
coercion, getItem resolution (~235 lines extracted).

dataframe_helpers.py: toPandas conversion, join constants/logic,
schema/dtypes mapping (~200 lines extracted).
