+++ sparkleframe/polarsdf/dataframe.py
@@ -9,25 +9,19 @@ import pyarrow as pa
 
 from sparkleframe.base.dataframe import DataFrame as BaseDataFrame
 from sparkleframe.polarsdf import types as sft
-from sparkleframe.polarsdf.column import Column, _polars_schema_for
-from sparkleframe.polarsdf.group import GroupedData
-from sparkleframe.polarsdf.types import (
-    BinaryType,
-    BooleanType,
-    ByteType,
-    DataType,
-    DateType,
-    DecimalType,
-    DoubleType,
-    FloatType,
-    IntegerType,
-    LongType,
-    ShortType,
-    StringType,
-    StructField,
-    StructType,
-    TimestampType,
+from sparkleframe.polarsdf.column import Column
+from sparkleframe.polarsdf.column_helpers import _polars_schema_for
+from sparkleframe.polarsdf.dataframe_helpers import (
+    PYSPARK_TO_POLARS_JOIN_MAP,
+    coalesce_outer_join_keys,
+    convert_pandas_value,
+    map_polars_dtype_to_spark_name,
+    perform_expression_join,
+    polars_dtype_to_spark_structfield,
+    resolve_equi_join_keys,
 )
+from sparkleframe.polarsdf.group import GroupedData
+from sparkleframe.polarsdf.types import DataType, StructField, StructType
 from sparkleframe.polarsdf.types_utils import _MapTypeUtils
 
 
@@ -376,80 +370,8 @@ class DataFrame(BaseDataFrame):
         Removes keys inside dicts where the value is None,
         but keeps the column and row structure intact.
         """
-        import math
-
-        import numpy as np
-        import pandas as pd
-
         df = self.df.to_arrow().to_pandas()
-
-        def is_null(x) -> bool:
-            try:
-                return pd.isna(x) and not isinstance(x, (str, bytes))
-            except Exception:
-                return False
-
-        def convert_number(x):
-            if isinstance(x, (np.integer,)):
-                return int(x)
-            if isinstance(x, int):
-                return x
-            if isinstance(x, (np.floating, float)):
-                if math.isnan(x):
-                    return None
-                return int(x) if float(x).is_integer() else float(x)
-            return x
-
-        # --- helpers to collapse Polars map layout(s) ---
-        def _is_kv(d) -> bool:
-            return isinstance(d, dict) and "key" in d and "value" in d
-
-        def _kv_list_to_dict(kv_list: list):
-            # [{"key":k,"value":v}, ...] -> {k: v}, skipping null values
-            out = {}
-            for item in kv_list:
-                k = convert(item["key"])
-                v = convert(item.get("value"))
-                if v is not None:
-                    out[k] = v
-            return out
-
-        def convert(val):
-            if is_null(val):
-                return None
-
-            # NumPy arrays -> list
-            if isinstance(val, np.ndarray):
-                return [convert(v) for v in val.tolist()]
-
-            # collapse map and array<map> encodings
-            if isinstance(val, list):
-                # Case 1: a single map encoded as a list of {"key","value"} dicts
-                if val and all(_is_kv(item) for item in val):
-                    return _kv_list_to_dict(val)
-
-                # Case 2: array<map<...>> encoded as list of kv-lists
-                if val and all(isinstance(item, list) and (not item or all(_is_kv(x) for x in item)) for item in val):
-                    return [_kv_list_to_dict(item) for item in val]
-
-                # General list -> recurse
-                return [convert(v) for v in val]
-
-            if isinstance(val, tuple):
-                return [convert(v) for v in val]
-
-            if isinstance(val, dict):
-                # Drop only keys whose converted value is None
-                out = {}
-                for k, v in val.items():
-                    cv = convert(v)
-                    if cv is not None:
-                        out[k] = cv
-                return out
-
-            return convert_number(val)
-
-        return df.map(convert)
+        return df.map(convert_pandas_value)
 
     def to_arrow(self) -> pa.Table:
         """
@@ -588,104 +510,29 @@ class DataFrame(BaseDataFrame):
                     break
             on = [n.to_native() if isinstance(n, Column) else n for n in on]
 
-        # Mapping of PySpark join types to Polars join types
-        PYSPARK_TO_POLARS_JOIN_MAP = {
-            "inner": "inner",
-            "cross": "cross",
-            "outer": "full",
-            "full": "full",
-            "fullouter": "full",
-            "full_outer": "full",
-            "left": "left",
-            "leftouter": "left",
-            "left_outer": "left",
-            "right": "right",
-            "rightouter": "right",
-            "right_outer": "right",
-            "semi": "semi",
-            "leftsemi": "semi",
-            "left_semi": "semi",
-            "anti": "anti",
-            "leftanti": "anti",
-            "left_anti": "anti",
-        }
-
         how = how.lower()
         if how not in PYSPARK_TO_POLARS_JOIN_MAP:
             raise ValueError(f"Unsupported join type: '{how}'")
 
         polars_join_type = PYSPARK_TO_POLARS_JOIN_MAP[how]
         suffix = "_" + str(uuid4()).replace("-", "")
-        # Only use cross-join + filter for a *boolean* join condition (e.g. col("a") == col("b")).
-        # Bare column refs (e.g. col("id")) are Spark equi-join keys, not predicates.
         use_expr_join = on is not None and len(on) == 1 and isinstance(on[0], pl.Expr) and not on[0].meta.is_column()
         if use_expr_join:
             if how not in {"inner", "left", "leftouter", "left_outer"}:
                 raise ValueError("Expression joins currently support only inner/left joins")
-            predicate = on[0]
-
-            left_with_idx = self.df.with_row_index("__sf_left_idx")
-            matched = left_with_idx.join(other.df, how="cross", suffix=suffix).filter(predicate)
-
-            if how in {"inner"}:
-                result = matched.drop("__sf_left_idx")
-            else:
-                matched_ids = matched.select("__sf_left_idx").unique()
-                unmatched = left_with_idx.join(matched_ids, on="__sf_left_idx", how="anti")
-
-                unmatched_right_cols = []
-                for right_col in other.df.columns:
-                    target_col = right_col if right_col not in self.df.columns else f"{right_col}{suffix}"
-                    unmatched_right_cols.append(pl.lit(None).alias(target_col))
-                unmatched = unmatched.with_columns(unmatched_right_cols).drop("__sf_left_idx")
-                result = pl.concat([matched.drop("__sf_left_idx"), unmatched], how="diagonal_relaxed")
+            result = perform_expression_join(self.df, other.df, on[0], how, suffix)
         else:
             equi_on: Union[None, str, list[str]] = on
             if on is not None:
-                key_names: list[str] = []
-                for x in on:
-                    if isinstance(x, str):
-                        key_names.append(x)
-                    elif isinstance(x, pl.Expr) and x.meta.is_column():
-                        key_names.append(x.meta.output_name().split(".")[-1])
-                    else:
-                        raise TypeError(f"Unsupported equi-join key expression: {x!r}")
-                equi_on = key_names[0] if len(key_names) == 1 else key_names
+                equi_on = resolve_equi_join_keys(on)
             result = self.df.join(other.df, on=equi_on, how=polars_join_type, suffix=suffix)
 
         if how == "outer":
-            """
-            Polars does not automatically coalesce join keys (e.g., id) in a full outer join because it retains both left and right keys explicitly, especially when:
-                * There are mismatches in the keys (e.g., id exists only on one side).
-                * It needs to distinguish between matching and non-matching keys.
+            result = coalesce_outer_join_keys(result, suffix, on_column_wrappers)
 
-            Why this happens?
-            Polars must preserve all information during a full (outer) join:
-                * If the key is missing on one side, it will still be included in the output, but with nulls on the missing side.
-                * Rather than overwrite or merge the column into one, it creates:
-                    - id from the left table
-                    - id_right (or similar suffix) from the right table
-
-            This ensures no loss of data or ambiguity, which is particularly important for:
-                * Asymmetric joins (like one-to-many).
-                * Duplicated key values or nulls.
-            """
-
-            for col in result.columns:
-                if col.endswith(suffix):
-
-                    # TODO: for some reason pyspark results from outer differs when `on_keys` are Column or str, wheter
-                    # the col is dropped or a coalesce happens
-                    if on_column_wrappers:
-                        result = result.drop(col)
-                    else:
-                        result = result.with_columns(
-                            pl.coalesce(col.replace(suffix, ""), col).alias(col.replace(suffix, ""))
-                        ).drop(col)
-
-        for col in result.columns:
-            if col.endswith(suffix):
-                result = result.rename({col: col.replace(suffix, "") + "_right"})
+        for col_name in result.columns:
+            if col_name.endswith(suffix):
+                result = result.rename({col_name: col_name.replace(suffix, "") + "_right"})
 
         return DataFrame(result)
 
@@ -699,134 +546,19 @@ class DataFrame(BaseDataFrame):
         Returns:
             List[Tuple[str, str]]: List of (column name, data type) pairs.
         """
-        POLARS_TO_PYSPARK_DTYPE_MAP = {
-            pl.Int8: "tinyint",
-            pl.Int16: "smallint",
-            pl.Int32: "int",
-            pl.Int64: "bigint",
-            pl.UInt8: "tinyint",
-            pl.UInt16: "smallint",
-            pl.UInt32: "int",
-            pl.UInt64: "bigint",
-            pl.Float32: "float",
-            pl.Float64: "double",
-            pl.Boolean: "boolean",
-            pl.Utf8: "string",
-            pl.Date: "date",
-            pl.Datetime: "timestamp",
-            pl.Time: "time",
-            pl.Duration: "interval",
-            pl.Object: "binary",
-            pl.List: "array",
-            pl.Struct: "struct",
-            pl.Decimal: "decimal",
-            pl.Binary: "binary",
-        }
-
-        def map_dtype(dtype: pl.DataType) -> str:
-            if isinstance(dtype, pl.Decimal):
-                return f"decimal({dtype.precision},{dtype.scale})"
-
-            if isinstance(dtype, pl.Struct):
-                # Recursively describe fields
-                fields_str = ",".join(f"{field.name}:{map_dtype(field.dtype)}" for field in dtype.fields)
-                return f"struct<{fields_str}>"
-
-            for polars_type, spark_type in POLARS_TO_PYSPARK_DTYPE_MAP.items():
-                if isinstance(dtype, polars_type):
-                    return spark_type
-
-            return str(dtype)
-
-        return [(col, map_dtype(dtype)) for col, dtype in self.df.schema.items()]
+        return [(col, map_polars_dtype_to_spark_name(dtype)) for col, dtype in self.df.schema.items()]
 
     @property
     def schema(self) -> StructType:
         """
         Mimics pyspark.sql.DataFrame.schema by returning the schema as a StructType.
         """
-
-        def _declared_type_for(col_name: str) -> Optional[DataType]:
-            if isinstance(self._schema, StructType):
-                for f in self._schema:
-                    if f.name == col_name:
-                        return f.dataType
-            return None
-
-        def _to_spark_datatype(dtype: pl.DataType, col_name: Optional[str] = None) -> DataType:
-            # --- NEW: detect native map layout (List(Struct["key","value"])) ---
-            if _MapTypeUtils.is_map_dtype(dtype):
-                key_dt_pl = dtype.inner.fields[0].dtype
-                val_dt_pl = dtype.inner.fields[1].dtype
-                key_dt = _to_spark_datatype(key_dt_pl)  # typically StringType()
-                val_dt = _to_spark_datatype(val_dt_pl)  # may itself be a Map/Array/Struct/etc.
-
-                # preserve declared valueContainsNull if we have it
-                value_contains_null = True
-                if col_name is not None:
-                    decl = _declared_type_for(col_name)
-                    if isinstance(decl, sft.MapType):
-                        value_contains_null = decl.valueContainsNull
-
-                return sft.MapType(key_dt, val_dt, valueContainsNull=value_contains_null)
-
-            # Decimals
-            if isinstance(dtype, pl.Decimal):
-                return DecimalType(dtype.precision, dtype.scale)
-
-            # Structs (recursive)
-            if isinstance(dtype, pl.Struct):
-                nested_fields = [StructField(f.name, _to_spark_datatype(f.dtype)) for f in dtype.fields]
-                return StructType(nested_fields)
-
-            # Arrays
-            if isinstance(dtype, pl.List):
-                elem_dtype = _to_spark_datatype(dtype.inner)
-                contains_null = True
-                if col_name is not None:
-                    decl = _declared_type_for(col_name)
-                    if isinstance(decl, sft.ArrayType):
-                        contains_null = decl.containsNull
-                return sft.ArrayType(elem_dtype, containsNull=contains_null)
-
-            # Scalars
-            POLARS_TO_SPARK = {
-                # Polars can infer Null dtype for all-null columns.
-                # Map it to StringType so downstream schema casts can still run.
-                pl.Null: StringType(),
-                pl.Utf8: StringType(),
-                pl.Int32: IntegerType(),
-                pl.UInt32: IntegerType(),
-                pl.Int64: LongType(),
-                pl.UInt64: LongType(),
-                pl.Float32: FloatType(),
-                pl.Float64: DoubleType(),
-                pl.Boolean: BooleanType(),
-                pl.Date: DateType(),
-                pl.Datetime: TimestampType(),
-                pl.Int8: ByteType(),
-                pl.UInt8: ByteType(),
-                pl.Int16: ShortType(),
-                pl.UInt16: ShortType(),
-                pl.Binary: BinaryType(),
-            }
-            for pl_type, spark_type in POLARS_TO_SPARK.items():
-                if isinstance(dtype, pl_type):
-                    return spark_type
-
-            raise TypeError(f"Unsupported dtype '{dtype}'")
-
-        def polars_dtype_to_spark_structfield(name: str, dtype: pl.DataType) -> StructField:
-            decl = _declared_type_for(name)
-            # Preserve declared ArrayType(MapType(...)) exactly as provided
-            if isinstance(decl, sft.ArrayType) and isinstance(decl.elementType, sft.MapType):
-                return StructField(name, decl)
-            # Preserve declared MapType exactly as provided
-            if isinstance(decl, sft.MapType):
-                return StructField(name, decl)
-            return StructField(name, _to_spark_datatype(dtype, col_name=name))
-
-        return StructType([polars_dtype_to_spark_structfield(name, dtype) for name, dtype in self.df.schema.items()])
+        return StructType(
+            [
+                polars_dtype_to_spark_structfield(name, dtype, declared_schema=self._schema)
+                for name, dtype in self.df.schema.items()
+            ]
+        )
 
     def sort(self, *cols: Union[str, Column, int, List[Union[str, Column, int]]]) -> DataFrame:
         """
